% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdata.R
\name{getDataset}
\alias{getDataset}
\title{getDataset}
\usage{
getDataset(conn, path)
}
\arguments{
\item{conn}{Object of type \code{h5pyd_connection} (from \code{initConn})}

\item{path}{Full path to the dataset to open}
}
\value{
object of type \code{Dataset}
}
\description{
Acquire a handle to a (read-only) h5pyd \code{Dataset} object
}
\examples{
D <- getDataset(conn, 'newassay001')
}
