% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdata.R
\name{getSubmatrix}
\alias{getSubmatrix}
\title{getSubmatrix}
\usage{
getSubmatrix(conn, dset, indices)
}
\arguments{
\item{conn}{Object of type \code{h5pyd_connection} (from \code{initConn})}

\item{dset}{Object of type \code{Dataset}}

\item{indices}{Character vector of slices. Must be in format 'start:stop:step' 
each integer-valued.  (R-conventions)}
}
\description{
Fetch data from \code{Dataset} object
}
\examples{
A <- getSubmatrix(conn, D, c('200:300:1', '400:500:1'))

}
