% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdata.R
\name{initConn}
\alias{initConn}
\title{initConn}
\usage{
initConn(domain, endpoint)
}
\arguments{
\item{domain}{H5File domain (see documentation)}

\item{endpoint}{endpoint}
}
\value{
object of type \code{h5pyd_connection}
}
\description{
Initialize connection to h5pyd server
}
\examples{
conn <- initConn(domain='/home/stvjc/tenx_full.h5', endpoint='http://52.4.181.237:5101') 

}
