% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdata.R
\name{initPython}
\alias{initPython}
\title{initPython}
\usage{
initPython(path = "")
}
\arguments{
\item{path}{The path to the Python executable to be used}
}
\description{
Initialize access to \code{reticulate} package and check
for installed modules \code{h5pyd} and \code{numpy}.
}
\examples{
initPython(path='/anaconda/bin/python')

}
