% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdata.R
\name{setSubmatrix}
\alias{setSubmatrix}
\title{setSubmatrix}
\usage{
setSubmatrix(dset, indices, mm)
}
\arguments{
\item{dset}{Object of type \code{Dataset}}

\item{indices}{Character vector of slices. Must be in format 'start:stop:step'}

\item{mm}{R array of data to store
each integer-valued.  (R-conventions)}
}
\description{
Store data to \code{Dataset} object
}
\examples{
setSubmatrix(D, c('1:2:1', '1:2:1', '1:2:1'), array(c(4, 14, 7, 17, 5, 15, 9, 19), dim=c(2, 2, 2)))

}
